Correct errors due to upstream changes (#63)

+++ src/test/scala/org/ode/engine/standardization/SoundHandler.scala
@@ -39,7 +39,7 @@ case class SoundHandler (
   sysBits: Int,
   wavBits: Int,
   sampleNumber: Int,
-  samplingRate: Double,
+  samplingRate: Float,
   chanNumber: Int
 ) {
